refactor(tracedecay): keep bench and test runtimes off lib
Default library builds no longer compile benchmark harnesses or
HostAdmissionTestRuntimeV1. The shipped tracedecay bench command
stays on a bench feature selected by the CLI production build.
