+++ crates/tracedecay-cli/Cargo.toml
@@ -71,7 +71,7 @@ hotpath-cpu = [
     "hotpath/hotpath-cpu",
 ]
 hotpath-mcp = ["hotpath", "hotpath/hotpath-mcp"]
-production = ["tracedecay/production"]
+production = ["tracedecay/production", "tracedecay/bench"]
 # Opt-in tiered graph storage; not part of `production`.
 graph-disk-tier = ["tracedecay/graph-disk-tier"]
 graph-tiered-storage = ["tracedecay/graph-tiered-storage"]
