+++ crates/tracedecay/benches/rmcp/benchmark.rs
