+++ crates/tracedecay/benches/session_temporal/root_relation_fixture.rs
