+++ crates/tracedecay/src/test_support/host_admission/profile_registry_test_support.rs
