+++ crates/tracedecay/src/test_support/mod.rs
@@ -1,0 +1,4 @@
+//! Test-only composition-root fixtures. Compiled under `cfg(test)` or
+//! `test-helpers`, never into a default or `production` library build.
+
+pub mod host_admission;
